requirements.txt
discord.py
hentai 
requests